ver.03

+++ 百度前端学院2017/task0003_个人任务管理系统_ver0.3/js/components/MenuList/index.js
@@ -10,7 +10,7 @@ class MenuList {
 		this.$container = $(opt.container)
 		this.$addFileBtn = $(opt.addFileBtn)
 		this.$exportBtn = $(opt.exportBtn) || null
-		this.$currentTarget = $('[data-folderId="-1"]')[0]
+		this.$currentTarget = null
 		this.guid =0
 
 		this.folders = []
@@ -43,16 +43,20 @@ class MenuList {
 
 		// addFileBtn 按钮的 添加file 事件
 		this.$addFileBtn.on('click', (e) => {
-			let folderId = this.$currentTarget.data('folderid')
-			let folderLevel = this.$currentTarget.data('folderlevel')
-			let folderInstance = this.getFolderById(folderId)
+			if(this.$currentTarget != null) {
+				let folderId = this.$currentTarget.data('folderid')
+				let folderLevel = this.$currentTarget.data('folderlevel')
+				let folderInstance = this.getFolderById(folderId)
 
-			if(folderInstance) {
-				//New File
-				let fileId = folderInstance.addFile().getLatestFile().fileId
+				if(folderInstance) {
+					//New File
+					let fileId = folderInstance.addFile().getLatestFile().fileId
 
-				pubsub.publish('updateCurrentFile',{folderId: folderId, fileId: fileId})
-				this.renderSubMenu( folderInstance )
+					pubsub.publish('updateCurrentFile',{folderId: folderId, fileId: fileId})
+					this.renderSubMenu( folderInstance )
+				}	
+			} else {
+				alert("新建任务前请至少选中一个文件夹")
 			}
 		})
 
@@ -65,7 +69,7 @@ class MenuList {
 			let fileId = $currentFile.data('fileid')
 			let folderInstance = this.getFolderById( folderId )
 
-			if(folderInstance) {
+			if(confirm("确认要删除该文件吗？") && folderInstance) {
 				folderInstance.deleteFile( fileId )
 				pubsub.publish('updateCurrentFile',{folderId: folderId, fileId: fileId})
 				this.renderSubMenu( folderInstance )
@@ -111,7 +115,34 @@ class MenuList {
 		this.folders.push( folder )
 	}
 
-	getData() {
+	deleteFolder( folderId )
+	{
+		let str = []
+
+		for(let i=0;i<this.folders.length;i++) {
+			if(this.folders[i].folderId == folderId) {
+				this.folders.splice(i,1)
+			}
+		}
+
+		for(let i=0;i<this.folders.length;i++) {
+			str += this.folders[i].element
+		}
+
+		this.$container.html( str )
+
+		//清空subMenu栏
+		if(this.$currentTarget != null) {
+			console.log(this.$currentTarget.data('folderid'), folderId)
+			if(this.$currentTarget.data('folderid') == folderId) {
+				$('.mid_main_list').html('')
+			}
+		}
+		
+	}
+
+	getData() 
+	{
 		let data = []
 		for(let i=0;i<this.folders.length;i++) {
 			if(this.folders[i]) {
@@ -120,6 +151,24 @@ class MenuList {
 		}
 
 		return data
+	}
+
+	getFilesHtml()
+	{
+		let subItems = []
+		let str = ''
+
+		for(let i=0;i<this.folders.length;i++) {
+			if(this.folders[i]) {
+				for(let j=0;j<this.folders[i].subItems.length;j++) {
+					this.folders[i].subItems[j] && subItems.push( this.folders[i].subItems[j] )
+				}
+			}
+		}
+
+		str = this.folders[0].getFilesHtml( subItems )
+
+		return str
 	}
 
 	loadData()
